Move group and collection components to their own directory.
