+++ scripts/components/collection/CollectionTabs.js
