+++ scripts/components/group/GroupTabs.js
