+++ scripts/containers/CollectionCreatePage.js
@@ -1,7 +1,7 @@
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 
-import CollectionCreate from "../components/CollectionCreate";
+import CollectionCreate from "../components/collection/CollectionCreate";
 import * as BucketActions from "../actions/bucket";
