+++ scripts/containers/GroupEditPage.js
@@ -1,7 +1,7 @@
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 
-import GroupEdit from "../components/GroupEdit";
+import GroupEdit from "../components/group/GroupEdit";
 import * as BucketActions from "../actions/bucket";
